internal/contour: move internal/route into internal/contour
Updates #454

Move the contents of internal/route into internal/contour. Deduplicate
test and helpers.
